Fix pointer arithmetic so adding a negative works
Previously, adding an integer to an address was done by converting the integer
to the type used to store the address. Because the address type is unsigned (and
the integer is signed), negative values would become positive, resulting in
addition overflow. This commit changes the implementation so that, when
adding a negative value, the integer is converted to a positive value and
subtraction is performed instead. A similar approach is used for address-integer
subtraction.