Add MOVE instruction
MOVE is quite powerful because it can take a value from an arbitrary
source and put it in an arbitrary destination. The format is:

MOVE <dest>, <src>

Silly operations are possible with it such as:

MOVE @, @

which is a noop because it just pops a value from the stack and then
pushes it back.